+++ csense-kotlin-jvm/src/main/kotlin/csense/kotlin/logger/LExtensions.kt
@@ -10,7 +10,7 @@ import kotlin.reflect.*
  * Uses the simple class name for the tag, otherwise just as the regular debug method
  * @see L.debug
  */
-inline fun L.debug(kClass: KClass<*>, message: String, exception: Throwable? = null) {
+inline fun LLogger.debug(kClass: KClass<*>, message: String, exception: Throwable? = null) {
     debug(kClass.java, message, exception)
 }
 
@@ -18,7 +18,7 @@ inline fun L.debug(kClass: KClass<*>, message: String, exception: Throwable? = n
  * Uses the simple class name for the tag, otherwise just as the regular debug method
  * @see L.warning
  */
-inline fun L.warning(kClass: KClass<*>, message: String, throwable: Throwable? = null) {
+inline fun LLogger.warning(kClass: KClass<*>, message: String, throwable: Throwable? = null) {
     warning(kClass.java, message, throwable)
 }
 
@@ -26,7 +26,7 @@ inline fun L.warning(kClass: KClass<*>, message: String, throwable: Throwable? =
  * Uses the simple class name for the tag, otherwise just as the regular debug method
  * @see L.error
  */
-inline fun L.error(kClass: KClass<*>, message: String, exception: Throwable? = null) {
+inline fun LLogger.error(kClass: KClass<*>, message: String, exception: Throwable? = null) {
     error(kClass.java, message, exception)
 }
 
@@ -34,7 +34,7 @@ inline fun L.error(kClass: KClass<*>, message: String, exception: Throwable? = n
  * Uses the simple class name for the tag, otherwise just as the regular debug method
  * @see L.logProd
  */
-inline fun L.logProd(kClass: KClass<*>, message: String, throwable: Throwable? = null) {
+inline fun LLogger.logProd(kClass: KClass<*>, message: String, throwable: Throwable? = null) {
     logProd(kClass.java, message, throwable)
 }
 
@@ -43,7 +43,7 @@ inline fun L.logProd(kClass: KClass<*>, message: String, throwable: Throwable? =
  * Uses the simple class name for the tag, otherwise just as the regular debug method
  * @see L.debug
  */
-inline fun L.debug(cls: Class<*>, message: String, exception: Throwable? = null) {
+inline fun LLogger.debug(cls: Class<*>, message: String, exception: Throwable? = null) {
     debug(cls.simpleName, message, exception)
 }
 
@@ -51,7 +51,7 @@ inline fun L.debug(cls: Class<*>, message: String, exception: Throwable? = null)
  * Uses the simple class name for the tag, otherwise just as the regular debug method
  * @see L.warning
  */
-inline fun L.warning(cls: Class<*>, message: String, throwable: Throwable? = null) {
+inline fun LLogger.warning(cls: Class<*>, message: String, throwable: Throwable? = null) {
     warning(cls.simpleName, message, throwable)
 }
 
@@ -59,7 +59,7 @@ inline fun L.warning(cls: Class<*>, message: String, throwable: Throwable? = nul
  * Uses the simple class name for the tag, otherwise just as the regular debug method
  * @see L.error
  */
-inline fun L.error(cls: Class<*>, message: String, exception: Throwable? = null) {
+inline fun LLogger.error(cls: Class<*>, message: String, exception: Throwable? = null) {
     error(cls.simpleName, message, exception)
 }
 
@@ -67,7 +67,7 @@ inline fun L.error(cls: Class<*>, message: String, exception: Throwable? = null)
  * Uses the simple class name for the tag, otherwise just as the regular debug method
  * @see L.logProd
  */
-inline fun L.logProd(cls: Class<*>, message: String, throwable: Throwable? = null) {
+inline fun LLogger.logProd(cls: Class<*>, message: String, throwable: Throwable? = null) {
     logProd(cls.simpleName, message, throwable)
 }