+++ changelog.md
@@ -11,7 +11,8 @@
 - cleaned up unused enum in Logger, as the level is now more concise with how its implemented 
 - Comparing is now "ItemComparing", since its between "items"
 - uniformed extensions on numbers (float, double, int, long, short, byte, now have +- the same extensions regarding positive, negative ect)
-  
+- extracted the L logging part into a class, such that it can be used in various ways, while still having a L object (also improves tests).
+    - also updated extensions.
 
 ## 0.0.14
 - fixed logging issue